BUGFIX #7 Refactoring players & matches models
Related to that issue, using referne paths solved the error when trying
to populate players but it has led to some new issues:

In order to populate the correct model ("Player" or "Team") a previous
fetch of the document to extract the prop storing this value was needed.
Having to do 2 reading petitions to the DB for each match view consult
is a lack in tems of performance.

Morover, when trying to get a doubles match view, it was kinda
difficulet to get the players info because a player populate over the
team populate was needed (and it's not know if possible).

Both issues has been solved by deleting Team model and refactoring players model so a team is an instance of players that has two referenced players.
Not sure if is worth it but this solution seems to fit alright.

On the other hand, an issue regarding compilation caused by module alias
has been solved by adding all the missing pathings in both files:
tsconfig and package.json