BUGFIX type of odds received
All the data related to odds is receives as strings and converted to
number before storing in DB (dont know if it's worth at all).

The value of an odd can be "-", representing the market is close and
it was being converted as NaN, crashing the server.

The odds data is not being saved as numbers anymore.