Add #16 Func + Fix #21
Bugfix #21: Add validations to check if the result is valid:

- Checking if the match have at least 2 sets

- If match is not doubles, the winner of the last set must have at least
  6 games (TAKE CARE: some ITF qualys have a super tie break on 3rd set
    as well. If so, a modification will have to be done)

If result is not correct, the match is saved without result and winner
and a document is stored in the ResultIssue collection.
At every call of the main endpoint, these documents are iterated looking
for updates.

Take care of UTR matches, most of them, doesn't have the final game of
the match.