Add logic to search and delete duplicate matches
Add logic inside saveAllPlayerMatches in order to identify those new
ended matches duplicated on DB and delete it.
The fact is that the API returns some upcoming matches and eventView
matches duplicated. Once they finish, the event views persists but the
ended Matches retrieved thaat for the player just have one of them.
The new funcionality identifies those matches that arent on the event
ended API data array, and deletes the missing match from DB only if it's
duplicated